Fixed problems with build on MacOS.  (Thanks for the patch...)
PointTest got broken somewhere along the way.  The problem is in Point::bearingDistanceTo where the old arithmetic there was written before the days of IEEE floating point and NAN.  In particular, the case where T = 0/0 is not as remote a possibility as nature would suggest.  The right answer here is T = 1.  IEEE decided it would be nan.

sigh.